Промежуточный этап рефакторинга
